Split top bar parity gates
