+++ docs/DECISIONS.md
@@ -285,3 +285,4 @@
 - Native sidebar command payloads are built through `QuillCodeSidebarCommandAdapter.swift`. SwiftUI row and bulk-action views should not construct `WorkspaceCommandSurface` values inline; selection toggles and bulk actions share the same adapter so command IDs, titles, categories, and enabled state stay consistent.
 - Static HTML browser snapshots live in `BrowserHTMLSnapshotBuilder`; `BrowserInspector` remains the URL/file/fetched-page adapter and delegates title/count/outline/snippet extraction to the focused builder.
 - Streaming shell command lifecycle lives in `ShellStreamingProcessRunner`; `ShellToolExecutor` remains the public shell facade for blocking, cancellable, and streaming entry points while the runner owns async stdout/stderr readers, timeout termination, cancellation, and final streamed `ToolResult` emission.
+- Top-bar parity gates live in `ParityTopBarGateTests`, not the broad `ParityGateTests` catch-all. Presentation semantics, lifecycle status labels, runtime/auth labels, model-catalog projection, and top-bar surface contract gates should stay together so top-bar UX can evolve without inflating the main architecture-gate file.
